pyproject.toml
[project]
name = "docparser"
version = "0.1.0"
description = "Universal document parser - converts PDF, images, Excel to canonical JSON with AI validation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "DocParser Team" }]

dependencies = [
    # API Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # OCR & Document Processing
    "google-cloud-vision>=3.5.0",
    "pymupdf>=1.23.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",

    # Structured Data
    "openpyxl>=3.1.0",
    "pandas>=2.1.0",
    "lxml>=5.0.0",

    # LLM Integration
    "openai>=1.10.0",
    "google-generativeai>=0.5.0",

    # Utilities
    "python-magic>=0.4.27",
    "httpx>=0.26.0",
    "pycountry>=23.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docparser"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
